+++ amfiteatr_core/src/agent/mod.rs
@@ -13,3 +13,5 @@ pub use traits::*;
 pub use trajectory::*;
 pub use generic::*;
 pub use info_set::*;
+
+