+++ amfiteatr_core/src/agent/traits/manual_control/action_command.rs
@@ -1,0 +1,4 @@
+
+pub enum ActionCommand<A: >{
+
+}