+++ amfiteatr_core/src/agent/trajectory.rs
@@ -83,7 +83,9 @@ impl<'a, DP: DomainParameters, S: InformationSet<DP>> AgentStepView<'a, DP, S>{
         self.action
     }
 
-    /// If action is legal returns `true`, else `false`
+    /// If action is legal returns `true`, else `false`. It says about `action` in `information_set`,
+    /// does not say anything about step initial information set, but says that next information set ([`late_information_set`])
+    /// is to be treated as illegal.
     pub fn is_action_legal(&self) -> bool{
         self.legality
     }